FIX:Wrapper,Recorder: record link on workspace
The previous implementation would record the link relatively to the cwd.
In some containers, this meant it was being recorded on '/'. This is
problematic, because the Jenkins process may not have write access to
that directory (for very valid reasons). Use the workspace variable to
make sure we can write to a valid location.